// src/app/zdslogic-ui-plugins/communicator/communicator-site/communicator-site-container/components/window-container/email-window/core/models/email-sent-file.model.ts
import { User } from 'app/zdslogic-ui-base/core/models/user.model';

export class EMailSentFile {

  id: string;

  userId: string;

  type: string;

  fromId: string;

  toId: string;

  fromText: string;

  replyText: string;

  toText: string;

  subject: string;

  message: string;

  dateSent: string;

  dateSeen: string;

  fromUser?: User = new User();

  toUser?: User = new User();

}